requirements.txt
Flask==0.10.1
Flask-Restless==0.13.1
Flask-SQLAlchemy==0.16
Jinja2==2.7
MarkupSafe==0.18
Werkzeug==0.9.2
gunicorn==17.5
itsdangerous==0.22
mimerender==0.5.4
python-dateutil==2.1
python-mimeparse==0.1.4
six==1.3.0
wsgiref==0.1.2
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
ecdsa==0.13
Fabric==1.11.1
kombu==3.0.35
paramiko==1.17.0
pycrypto==2.6.1
python-digitalocean==1.8.1
pytz==2016.4
redis==2.10.5
requests==2.10.0
SQLAlchemy==1.0.12
wheel==0.24.0
